apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

SQLAlchemy==2.0.34
psycopg[binary]==3.2.13
alembic==1.13.2

pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Auth
pyjwt[crypto]==2.9.0
cryptography==43.0.1

# Email
resend==2.1.0
